Migrate getting filename to utils.ts
